requirements.txt
django
python-dotenv
psycopg2
django-prometheus
waitress
whitenoise
django-basicauth